Add SurfaceRZFourier class for surface representation and interpolation; include initial debugging prints
